+++ app/RpsCompetition/Frontend/WpseoHelper.php
@@ -5,6 +5,7 @@ use RpsCompetition\Common\Helper as CommonHelper;
 use RpsCompetition\Db\QueryCompetitions;
 use RpsCompetition\Db\QueryMiscellaneous;
 use RpsCompetition\Db\RpsDb;
+use RpsCompetition\Photo\Helper as PhotoHelper;
 use RpsCompetition\Settings;
 
 if (!class_exists('AVH_RPS_Client')) {
@@ -20,19 +21,25 @@ if (!class_exists('AVH_RPS_Client')) {
  */
 class WpseoHelper
 {
+    private $photo_helper;
+    private $query_miscellaneous;
     private $rpsdb;
     private $settings;
 
     /**
      * Constructor
      *
-     * @param Settings $settings
-     * @param RpsDb    $rpsdb
+     * @param Settings           $settings
+     * @param RpsDb              $rpsdb
+     * @param QueryMiscellaneous $query_miscellaneous
+     * @param PhotoHelper        $photo_helper
      */
-    public function __construct(Settings $settings, RpsDb $rpsdb)
+    public function __construct(Settings $settings, RpsDb $rpsdb, QueryMiscellaneous $query_miscellaneous, PhotoHelper $photo_helper)
     {
         $this->settings = $settings;
         $this->rpsdb = $rpsdb;
+        $this->query_miscellaneous = $query_miscellaneous;
+        $this->photo_helper = $photo_helper;
     }
 
     /**
@@ -51,7 +58,7 @@ class WpseoHelper
     {
         $options = get_option('avh-rps');
         $url = get_permalink($options['monthly_entries_post_id']);
-        $this->buildWpseoSitemap($url);
+        $this->buildWpseoSitemap($url, true);
         exit();
     }
 
@@ -87,6 +94,44 @@ class WpseoHelper
     }
 
     /**
+     * Change the Sitemap data for the front page.
+     * As WordPress SEO by Yoast does not change the data for the front page, we do this ourselves.
+     * We set the priority to 100%, the Change Frequency to Daily, and set the last modified date
+     * by getting the articles on the front page and using the most recent modified date of that post.
+     * If we don't do this, the entry will take the modified date from the actual page, which never changes.
+     *
+     * @param array  $data
+     * @param string $type
+     * @param object $current_post
+     *
+     * @return array
+     */
+    public function filterSitemapEntry($data, $type, $current_post)
+    {
+        if ($current_post->ID == get_option('page_on_front')) {
+            $data['pri'] = 1;
+            $data['chf'] = 'daily';
+
+            $post_date_modified = [];
+            // Setup query for sticky posts.
+            $sticky = get_option('sticky_posts');
+            // Get the query for articles marked for Magazine Excerpts marked per post and in the selected categories.
+            $queries = rps_suffusion_get_mag_section_queries(array('meta_check_field' => 'suf_magazine_excerpt', 'category_prefix' => 'suf_mag_excerpt_categories', 'to_skip' => $sticky));
+            foreach ($queries as $query) {
+                $posts = get_posts($query->query);
+                foreach ($posts as $p) {
+                    $post_date_modified[] = $p->post_modified;
+                }
+            }
+            rsort($post_date_modified);
+            $date_modified = new \DateTime(reset($post_date_modified));
+            $data['mod'] = $date_modified->format('c');
+        }
+
+        return $data;
+    }
+
+    /**
      * Filter for the title of pages.
      *
      * @param array $title_array
@@ -107,7 +152,7 @@ class WpseoHelper
             $competition = current($competitions);
 
             $new_title_array = array();
-            $new_title_array[] = $post->post_title . ' for the theme "' .$competition->Theme . '" on ' . $date_text;
+            $new_title_array[] = $post->post_title . ' for the theme "' . $competition->Theme . '" on ' . $date_text;
             $title_array = $new_title_array;
         }
 
@@ -225,14 +270,40 @@ class WpseoHelper
     {
         $query_competitions = new QueryCompetitions($this->settings, $this->rpsdb);
 
+        $all_competitions = $query_competitions->getScoredCompetitions('1970-01-01', '2200-01-01');
+        $years = [];
+        $old_year = 0;
+        $old_mod_date = 0;
+        foreach ($all_competitions as $competition) {
+            $date = new \DateTime($competition->Competition_Date);
+            $year = $date->format('Y');
+
+            $date_modified = new \DateTime($competition->Date_Modified, new \DateTimeZone(get_option('timezone_string')));
+            $mod_date = $date_modified->format('U');
+            $last_modified_date = $date_modified->format('c');
+
+            if ($year != $old_year) {
+                $old_mod_date = 0;
+                $old_year = $year;
+            }
+            if ($mod_date > $old_mod_date) {
+                $old_mod_date = $mod_date;
+                $last_modified_date = $date_modified->format('c');
+            }
+            $years[$year] = $last_modified_date;
+        }
+
         $last_scored = $query_competitions->query(array('where' => 'Scored="Y"', 'orderby' => 'Date_Modified', 'order' => 'DESC', 'number' => 1));
         $date = new \DateTime($last_scored->Date_Modified);
 
         $sitemap = '';
-        $sitemap .= '<sitemap>' . "\n";
-        $sitemap .= '<loc>' . wpseo_xml_sitemaps_base_url('competition-entries') . '-sitemap.xml</loc>' . "\n";
-        $sitemap .= '<lastmod>' . htmlspecialchars($date->format('c')) . '</lastmod>' . "\n";
-        $sitemap .= '</sitemap>' . "\n";
+
+        foreach ($years as $year => $lastmod) {
+            $sitemap .= '<sitemap>' . "\n";
+            $sitemap .= '<loc>' . wpseo_xml_sitemaps_base_url('competition-entries') . '-sitemap' . $year . '.xml</loc>' . "\n";
+            $sitemap .= '<lastmod>' . htmlspecialchars($lastmod) . '</lastmod>' . "\n";
+            $sitemap .= '</sitemap>' . "\n";
+        }
         $sitemap .= '<sitemap>' . "\n";
         $sitemap .= '<loc>' . wpseo_xml_sitemaps_base_url('competition-winners') . '-sitemap.xml</loc>' . "\n";
         $sitemap .= '<lastmod>' . htmlspecialchars($date->format('c')) . '</lastmod>' . "\n";
@@ -269,11 +340,22 @@ class WpseoHelper
      * Build actual sitemap for Competition Entries or Competition Winners
      *
      * @param string $url
+     * @param bool   $include_images
      */
-    private function buildWpseoSitemap($url)
+    private function buildWpseoSitemap($url, $include_images = false)
     {
+        $n = get_query_var('sitemap_n');
+        if (is_scalar($n) && intval($n) > 0) {
+            $n = intval($n);
+        }
+        $data = [];
+        $date = new \DateTime();
+        $date->setDate($n, 1, 1);
+        $start_date = $date->format('Y-m-d');
+        $date->setDate($n, 12, 31);
+        $end_date = $date->format('Y-m-d');
         $query_competitions = new QueryCompetitions($this->settings, $this->rpsdb);
-        $scored_competitions = $query_competitions->getScoredCompetitions('1970-01-01', '2200-01-01');
+        $scored_competitions = $query_competitions->getScoredCompetitions($start_date, $end_date);
 
         $old_mod_date = 0;
         $old_key = 0;
@@ -299,9 +381,28 @@ class WpseoHelper
             $sitemap_data[$key] = array(
                 'loc' => $location,
                 'pri' => 0.8,
-                'chf' => 'monthly',
+                'chf' => 'yearly',
                 'mod' => $last_modified_date,
             );
+
+            if ($include_images) {
+                $entries = $this->query_miscellaneous->getAllEntries($competition_date->format('Y-m-d'));
+                $data['images'] = array();
+                if (is_array($entries)) {
+                    // Iterate through all the award winners and display each thumbnail in a grid
+                    /** @var QueryEntries $entry */
+                    foreach ($entries as $record) {
+                        $user_info = get_userdata($record->Member_ID);
+                        $title = $record->Title;
+                        $last_name = $user_info->user_lastname;
+                        $first_name = $user_info->user_firstname;
+                        $data['images'][]['loc'] = $this->photo_helper->getThumbnailUrl($record->Server_File_Name, '800');
+                        $data['images'][]['title'] = $title;
+                        $data['images'][]['caption'] = $title . ' Credit: ' . $first_name . ' ' . $last_name;
+                    }
+                }
+                $sitemap_data[$key]['images'] = $data['images'];
+            }
         }
         $this->outputWpseoSitemap($sitemap_data);
     }
@@ -316,15 +417,32 @@ class WpseoHelper
         $output = '';
         foreach ($sitemap_data as $data) {
             $output .= "\t<url>\n";
-            $output .= "\t\t<loc>" . $data['loc'] . "</loc>\n";
+            $output .= "\t\t<loc>" . htmlspecialchars($data['loc']) . "</loc>\n";
             $output .= "\t\t<lastmod>" . $data['mod'] . "</lastmod>\n";
             $output .= "\t\t<changefreq>" . $data['chf'] . "</changefreq>\n";
             $output .= "\t\t<priority>" . $data['pri'] . "</priority>\n";
+
+            if (isset($data['images']) && (is_array($data['images']) && $data['images'] !== array())) {
+                foreach ($data['images'] as $img) {
+                    if (!isset($img['loc']) || empty($img['loc'])) {
+                        continue;
+                    }
+                    $output .= "\t\t<image:image>\n";
+                    $output .= "\t\t\t<image:loc>" . esc_html($img['loc']) . "</image:loc>\n";
+                    if (isset($img['title']) && !empty($img['title'])) {
+                        $output .= "\t\t\t<image:title><![CDATA[" . _wp_specialchars(html_entity_decode($img['title'], ENT_QUOTES, esc_attr(get_bloginfo('charset')))) . "]]></image:title>\n";
+                    }
+                    if (isset($img['caption']) && !empty($img['caption'])) {
+                        $output .= "\t\t\t<image:caption><![CDATA[" . _wp_specialchars(html_entity_decode($img['caption'], ENT_QUOTES, esc_attr(get_bloginfo('charset')))) . "]]></image:caption>\n";
+                    }
+                    $output .= "\t\t</image:image>\n";
+                }
+            }
             $output .= "\t</url>\n";
         }
 
         $sitemap = '<urlset xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" ';
-        $sitemap .= 'xsi:schemaLocation="http://www.sitemaps.org/schemas/sitemap/0.9 http://www.sitemaps.org/schemas/sitemap/0.9/sitemap.xsd" ';
+        $sitemap .= 'xsi:schemaLocation="http://www.sitemaps.org/schemas/sitemap/0.9 http://www.sitemaps.org/schemas/sitemap/0.9/sitemap.xsd" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1" ';
         $sitemap .= 'xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">' . "\n";
         $sitemap .= $output;
         $sitemap .= '</urlset>';